requirements.txt
numpy
matplotlib
pandas
statsmodels
PyYAML
python-dotenv
requests
